% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaitSymmetry.R
\docType{data}
\name{GaitSymmetry}
\alias{GaitSymmetry}
\title{GaitSymmetry data}
\format{
An object of class \code{SPMBayes}
}
\usage{
data(GaitSymmetry)
}
\arguments{
\item{GaitSymmetry}{The data for GaitSymmetry}
}
\description{
Data from Single subject dataset (healthy male, 28
years, 83 kg, 178 cm) of left and right leg knee
flexion angles during 99 gait cycles on a dual-belt
tredmill at constant speed of 4.5 km/h, time
normalized to 101 time samples. Gait kinematics
were recorded with a 6-camera VICON system at 250Hz.
}
\examples{
head(data(GaitSymmetry))
}
\keyword{datasets}
