% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantarArchAngle.R
\docType{data}
\name{PlantarArchAngle}
\alias{PlantarArchAngle}
\title{PlantarArchAngle data}
\format{
An object of class \code{SPMBayes}
}
\usage{
data(PlantarArchAngle)
}
\arguments{
\item{PlantarArchAngle}{The data for PlantarArchAngle}
}
\description{
Dataset of 2 x n = 10 experimental time series of the plantar arch
angle of the foot at 101 time samples each
}
\examples{
head(data(PlantarArchAngle))
}
\keyword{datasets}
