% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedTwoLocalMax.R
\docType{data}
\name{SimulatedTwoLocalMax}
\alias{SimulatedTwoLocalMax}
\title{SimulatedTwoLocalMax data}
\format{
An object of class \code{SPMBayes}
}
\usage{
data(SimulatedTwoLocalMax)
}
\arguments{
\item{SimulatedTwoLocalMax}{The data for SimulatedTwoLocalMax}
}
\description{
Dataset of 2 x n = 10 experimental time series of the plantar arch
angle of the foot at 101 time samples each
}
\examples{
head(data(SimulatedTwoLocalMax))
}
\keyword{datasets}
