% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSPM.R
\name{SPMBayes}
\alias{SPMBayes}
\alias{plot.SPMBayes}
\title{-SPMBayes-}
\usage{
SPMBayes(
  Y,
  group,
  time,
  threshold = 0.95,
  Q.threshold = 0.05,
  Hypothesis = "alt"
)

\method{plot}{SPMBayes}(x, summary = TRUE, ...)
}
\arguments{
\item{Y}{(J x T) array specifying observations for T time points.}

\item{group}{(J x 1) factor vector specify the groups to which each observation belongs.}

\item{time}{(T x 1)  vector specify the times (\% stance)}

\item{threshold}{scalar specifying the posterior probability threshold.}

\item{Q.threshold}{scalar specifying a less conservative threshold using the Q-value}

\item{Hypothesis}{between "alt" and "null" specifying which alternative we are interested in.}

\item{x}{Summary object of class \code{SPMBayes}.}

\item{summary}{Logical stating whether to only give summary plot with mean and standad deviation}

\item{...}{Further arguments passed to ggplot2}
}
\description{
An R-package for conducting Bayesian SPM Analysis
}
\examples{
Y = mvtnorm::rmvnorm(40,mean=rep(0,50))
grp=rep(c(1,2,3,4),each=10)
time=1:50
out <- SPMBayes(Y,group=grp, time, threshold=0.95, Q.threshold=0.05, Hypothesis="alt")
plot(x=out, summary = TRUE) ## Get summary plot
plot(x=out, summary = FALSE) ## Get SPM Map plot

}
